Refactor to avoid adding to project_allocation_mapper
